% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Create a pp module.}
\usage{
pp(..., subclass = NULL)
}
\arguments{
\item{...}{Any object the module should contain.}

\item{subclass}{A character value. If `NULL`, the module will have class 'pp'.
If not `NULL`, the module will have subclass of 'pp'.}
}
\description{
Create a pp module.
}
