% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_instant.R
\name{pp_instant}
\alias{pp_instant}
\title{Create a pp module which evaluates expression specified in its argument.}
\usage{
pp_instant(expr, sub_object = NULL, desc = NULL, data = NULL,
  standby = TRUE)
}
\arguments{
\item{expr}{Expression.}

\item{sub_object}{Any object necessary to evaluate `expr`.}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which evaluates expression specified in its argument.
}
