% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_na_impute.R
\name{pp_na_impute}
\alias{pp_na_impute}
\title{Create a pp module which imputes missing data.}
\usage{
pp_na_impute(num_cols, cat_cols, num_method = "mean",
  cat_method = "frequent", desc = NULL, data = NULL,
  standby = TRUE)
}
\arguments{
\item{num_cols}{A character vector. Specify the names of numerical features.}

\item{cat_cols}{A character vector. Specify the names of categorical features.}

\item{num_method}{A character. Method for imputing numerical features.
Choose one of ("mean", "median", "frequent").}

\item{cat_method}{A character. Method for imputing categorical features.
Only "frequent" is available.}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which imputes missing data.
}
