% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_normalize.R
\name{pp_normalize}
\alias{pp_normalize}
\title{Create a pp module which normalizes numerical features.}
\usage{
pp_normalize(cols = NULL, method = c("mean-sd", "zero-one"),
  desc = NULL, data = NULL, standby = TRUE)
}
\arguments{
\item{cols}{A character vector. Specify the names of features to normalize.}

\item{method}{A character, one of ("mean-sd", "zero-one").
If `"mean-sd"`, subtract by mean and then divide by standard deviation.
If `"zero-one"`, fit range into [0, 1].}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which normalizes numerical features.
}
