% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_pca.R
\name{pp_pca}
\alias{pp_pca}
\title{Create a pp module which converts numerical features to principal components.}
\usage{
pp_pca(cols, threshold = 0.8, desc = NULL, data = NULL,
  standby = TRUE)
}
\arguments{
\item{cols}{A character vector. Specify the names of features to convert.}

\item{threshold}{A numeric value in [0, 1).}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which converts numerical features to principal components.
}
