% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_remove.R
\name{pp_remove}
\alias{pp_remove}
\title{Create a pp module which converts numerical features to principal components.}
\usage{
pp_remove(cols, desc = NULL, data = NULL, standby = TRUE)
}
\arguments{
\item{cols}{A character vector. Specify the names of features to remove.}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which converts numerical features to principal components.
}
