% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_sequential.R
\name{pp_sequential}
\alias{pp_sequential}
\title{Create pp_sequential module.}
\usage{
pp_sequential(..., data, pplist = NULL)
}
\arguments{
\item{...}{pp modules.}

\item{data}{A data frame. It should be a training data.}

\item{pplist}{A list contains pp modules.}
}
\description{
Create pp_sequential module.
}
