% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_transform.R
\name{pp_transform}
\alias{pp_transform}
\title{Create a pp module which transform features.}
\usage{
pp_transform(cols, fun, desc = NULL, data = NULL, standby = TRUE)
}
\arguments{
\item{cols}{A character vector. Specify the names of features to convert.}

\item{fun}{Transformation function such as `log`, `sqrt`, etc.}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which transform features.
}
